+++ tests/test.c
@@ -2743,8 +2743,11 @@ test_consume_txn_msgs_easy (const char *group_id, const char *topic,
 /**
  * @brief Waits for up to \p timeout_ms for consumer to receive assignment.
  *        If no assignment received without the timeout the test fails.
+ *
+ * @warning This method will poll the consumer and might thus read messages.
+ *          Set \p do_poll to false to use a sleep rather than poll.
  */
-void test_consumer_wait_assignment (rd_kafka_t *rk) {
+void test_consumer_wait_assignment (rd_kafka_t *rk, rd_bool_t do_poll) {
         rd_kafka_topic_partition_list_t *assignment = NULL;
         int i;
 
@@ -2760,7 +2763,10 @@ void test_consumer_wait_assignment (rd_kafka_t *rk) {
 
                 rd_kafka_topic_partition_list_destroy(assignment);
 
-                test_consumer_poll_once(rk, NULL, 1000);
+                if (do_poll)
+                        test_consumer_poll_once(rk, NULL, 1000);
+                else
+                        rd_usleep(1000*1000, NULL);
         }
 
         TEST_SAY("%s: Assignment (%d partition(s)): ",
@@ -4023,19 +4029,28 @@ int test_consumer_poll_once (rd_kafka_t *rk, test_msgver_t *mv, int timeout_ms){
 }
 
 
-int test_consumer_poll (const char *what, rd_kafka_t *rk, uint64_t testid,
-                        int exp_eof_cnt, int exp_msg_base, int exp_cnt,
-			test_msgver_t *mv) {
+/**
+ * @param exact Require exact exp_eof_cnt (unless -1) and exp_cnt (unless -1).
+ *              If false: poll until either one is reached.
+ */
+int test_consumer_poll_exact (const char *what, rd_kafka_t *rk, uint64_t testid,
+                              int exp_eof_cnt, int exp_msg_base, int exp_cnt,
+                              rd_bool_t exact, test_msgver_t *mv) {
         int eof_cnt = 0;
         int cnt = 0;
         test_timing_t t_cons;
 
-        TEST_SAY("%s: consume %d messages\n", what, exp_cnt);
+        TEST_SAY("%s: consume %s%d messages\n", what,
+                 exact ? "exactly ": "", exp_cnt);
 
         TIMING_START(&t_cons, "CONSUME");
 
-        while ((exp_eof_cnt <= 0 || eof_cnt < exp_eof_cnt) &&
-               (exp_cnt <= 0 || cnt < exp_cnt)) {
+        while ((!exact &&
+                ((exp_eof_cnt <= 0 || eof_cnt < exp_eof_cnt) &&
+                 (exp_cnt <= 0 || cnt < exp_cnt))) ||
+               (exact &&
+                (eof_cnt < exp_eof_cnt ||
+                 cnt < exp_cnt))) {
                 rd_kafka_message_t *rkmessage;
 
                 rkmessage = rd_kafka_consumer_poll(rk, tmout_multip(10*1000));
@@ -4067,6 +4082,13 @@ int test_consumer_poll (const char *what, rd_kafka_t *rk, uint64_t testid,
                                  rd_kafka_message_errstr(rkmessage));
 
                 } else {
+                        TEST_SAYL(4, "%s: consumed message on %s [%"PRId32"] "
+                                  "at offset %"PRId64"\n",
+                                  what,
+                                  rd_kafka_topic_name(rkmessage->rkt),
+                                  rkmessage->partition,
+                                  rkmessage->offset);
+
 			if (!mv || test_msgver_add_msg(rk, mv, rkmessage))
 				cnt++;
                 }
@@ -4079,12 +4101,28 @@ int test_consumer_poll (const char *what, rd_kafka_t *rk, uint64_t testid,
         TEST_SAY("%s: consumed %d/%d messages (%d/%d EOFs)\n",
                  what, cnt, exp_cnt, eof_cnt, exp_eof_cnt);
 
+        TEST_ASSERT(!exact ||
+                    ((exp_cnt == -1 || exp_cnt == cnt) &&
+                     (exp_eof_cnt == -1 || exp_eof_cnt == eof_cnt)),
+                    "%s: mismatch between exact expected counts and actual: "
+                    "%d/%d EOFs, %d/%d msgs",
+                    what, eof_cnt, exp_eof_cnt, cnt, exp_cnt);
+
         if (exp_cnt == 0)
                 TEST_ASSERT(cnt == 0 && eof_cnt == exp_eof_cnt,
                             "%s: expected no messages and %d EOFs: "
                             "got %d messages and %d EOFs",
                             what, exp_eof_cnt, cnt, eof_cnt);
         return cnt;
+}
+
+
+int test_consumer_poll (const char *what, rd_kafka_t *rk, uint64_t testid,
+                        int exp_eof_cnt, int exp_msg_base, int exp_cnt,
+                        test_msgver_t *mv) {
+        return test_consumer_poll_exact(what, rk, testid,
+                                        exp_eof_cnt, exp_msg_base, exp_cnt,
+                                        rd_false/*not exact */, mv);
 }
 
 void test_consumer_close (rd_kafka_t *rk) {