+++ tests/test.h
@@ -527,11 +527,14 @@ void test_consumer_poll_no_msgs (const char *what, rd_kafka_t *rk,
 void test_consumer_poll_expect_err (rd_kafka_t *rk, uint64_t testid,
                                     int timeout_ms, rd_kafka_resp_err_t err);
 int test_consumer_poll_once (rd_kafka_t *rk, test_msgver_t *mv, int timeout_ms);
+int test_consumer_poll_exact (const char *what, rd_kafka_t *rk, uint64_t testid,
+                              int exp_eof_cnt, int exp_msg_base, int exp_cnt,
+                              rd_bool_t exact, test_msgver_t *mv);
 int test_consumer_poll (const char *what, rd_kafka_t *rk, uint64_t testid,
                         int exp_eof_cnt, int exp_msg_base, int exp_cnt,
 			test_msgver_t *mv);
 
-void test_consumer_wait_assignment (rd_kafka_t *rk);
+void test_consumer_wait_assignment (rd_kafka_t *rk, rd_bool_t do_poll);
 void test_consumer_assign (const char *what, rd_kafka_t *rk,
                            rd_kafka_topic_partition_list_t *parts);
 void test_consumer_incremental_assign (const char *what, rd_kafka_t *rk,